SubprocessInput compilation moved
